REFACTOR: 모바일 반응형 레이아웃 적용
