+++ src/components/common/Footer/Footer.scss
@@ -1,3 +1,5 @@
+@use '../../../styles/breakpoints' as *;
+
 .footer {
   width: 100%;
   background-color: #1c1c1f;
@@ -15,7 +17,11 @@
   padding: 24px 0;
   border-top: 1px solid #424254;
 
-  @media (max-width: 768px) {
+  @include mobile {
+    padding: 24px 16px;
+  }
+
+  @include small {
     flex-direction: column;
     gap: 32px;
   }
@@ -28,7 +34,7 @@
   gap: 48px;
   margin-bottom: 48px;
 
-  @media (max-width: 768px) {
+  @include small {
     flex-direction: column;
     gap: 32px;
   }
@@ -94,7 +100,7 @@
   height: 24px;
   flex-wrap: wrap;
 
-  @media (max-width: 768px) {
+  @include small {
     justify-content: flex-start;
   }
