+++ src/components/home/FeaturesSection/FeaturesSection.scss
@@ -1,3 +1,5 @@
+@use '../../../styles/breakpoints' as *;
+
 .features-section {
   padding-top: 4px;
 }
@@ -6,14 +8,9 @@
   display: flex;
   justify-content: space-between;
 
-  @media (max-width: 1024px) {
-    grid-template-columns: repeat(2, 1fr);
-    gap: 30px;
-  }
-
-  @media (max-width: 768px) {
-    grid-template-columns: 1fr;
-    gap: 20px;
+  @include mobile {
+    flex-direction: column;
+    gap: 4px;
   }
 }
 
@@ -21,8 +18,13 @@
   width: 49.8%;
   height: 550px;
   background-color: #31313c;
-  border-radius: 4px;
+  border-radius: 4px 0 4px 4px;
   overflow: hidden;
+
+  @include mobile {
+    width: 100%;
+    border-radius: 4px;
+  }
 }
 
 .feature-item1-title {
@@ -63,7 +65,6 @@
     background-color: #282830;
   }
 
-  // 지원하지 않는 플랫폼
   &.not-supported {
     opacity: 0.3;
 
@@ -110,7 +111,6 @@
     }
   }
 
-  // 지원하지 않는 플랫폼에서는 platform-name 호버 효과 제거
   &.not-supported {
     .platform-name {
       cursor: default;
@@ -122,15 +122,21 @@
     }
   }
 }
+
 .tool-section {
   display: flex;
   flex-direction: column;
   align-items: center;
   width: 49.8%;
   background-color: #31313c;
-  border-radius: 4px;
+  border-radius: 0 4px 4px 4px;
   color: #7b7a8e;
   font-family: "Inter", sans-serif;
+
+  @include mobile {
+    display: none;
+  }
+
   h3 {
     color: #ffffff;
     font-size: 20px;
@@ -145,13 +151,11 @@
     margin-bottom: 32px;
   }
 
-  // 호버 시 이미지 회전 효과
   &:hover .tool-image {
     animation: spin 4s linear infinite;
   }
 }
 
-// 회전 애니메이션 정의
 @keyframes spin {
   from {
     transform: rotate(0deg);
@@ -165,7 +169,6 @@
   margin-top: 120px;
   width: 80px;
   height: 80px;
-  // 부드러운 전환을 위한 기본 설정
   transition: transform 0.3s ease;
 }
 
@@ -174,7 +177,7 @@
   justify-content: center;
   gap: 24px;
 
-  @media (max-width: 768px) {
+  @include small {
     flex-direction: column;
     align-items: center;
     gap: 16px;
