+++ src/components/result/ResultPageSearch/ResultSearchBar.scss
@@ -1,62 +1,21 @@
-.result-header-container {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  width: 1040px;
-  height: 56px;
-}
-.result-logo {
-  display: flex;
-  align-items: center;
-  height: 56px;
-  font-family: "Noto Sans KR", "Pretendard", sans-serif;
-  font-weight: 900;
-  font-size: 20px;
-  color: #ffffff;
-  padding: 0;
-  cursor: pointer;
-  transition: color 0.1s ease;
-}
-.result-header-right {
-  display: flex;
-  justify-content: flex-end;
-  gap: 20px;
-  width: 200px;
-  height: 56px;
-  font-size: 14px;
-  font-weight: 700;
-  color: #b3cdff;
-}
-.result-header-right-item {
-  height: 56px;
-  background: none;
-  border: none;
-  color: #b3cdff;
-  font-size: 14px;
-  font-weight: 700;
-  font-family: inherit;
-  border-radius: 0;
-  cursor: pointer;
-  padding: 0;
-  transition: color 0.1s ease;
-  &:hover {
-    color: #ffffff;
-  }
-}
-/* 검색 폼 컨테이너 스타일 */
+@use '../../../styles/breakpoints' as *;
+
 .result-page-search-form-container {
-  margin: 0 auto;
   width: 100%;
-  max-width: 700px;
-  padding-left: 60px;
+  max-width: 600px;
   position: relative;
+
+
+  @include mobile {
+    max-width: 100%;
+  }
 
   .search-form {
     display: flex;
-    height: 36px;
+    height: 32px;
     align-items: stretch;
     background-color: #fff;
-    border-radius: 4px;
+    border-radius: 2px;
     border: none;
     position: relative;
 
@@ -98,18 +57,54 @@
     }
   }
 
+  .region-inner {
+    display: flex;
+    align-items: stretch;
+    text-align: left;
+
+    & > div {
+      display: flex;
+    }
+  }
+
   .url-input {
     color: #000000;
+    height: 20px;
+    width: 100%;
+    box-sizing: border-box;
+    padding-left: 12px;
+    background-color: transparent;
+    border: none;
+    font-size: 14px;
+    position: relative;
+    z-index: 2;
+
+    &:focus {
+      outline: none;
+    }
+
+    &::placeholder {
+      color: #9e9eb1;
+    }
   }
 
   .platform-button {
+    position: relative;
     width: 110px;
     padding-left: 20px;
     padding-right: 20px;
-    background-color: transparent;
+    background-color: rgb(236, 242, 255);
     border-radius: 4px 0 0 4px;
     color: #4270d3;
+    border: none;
     border-right: 1px solid rgb(66, 112, 211, 0.8);
+    cursor: pointer;
+
+    @include mobile {
+      width: 100px;
+      padding-left: 12px;
+      padding-right: 12px;
+    }
 
     &::after {
       content: "";
@@ -129,16 +124,47 @@
     }
   }
 
+  .hidden {
+    display: none;
+  }
+
+  .mobile-only {
+    display: none;
+  }
+
+  .input-container {
+    position: relative;
+    flex: 1;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+  }
+
+  .sr-only {
+    position: absolute;
+    width: 1px;
+    height: 1px;
+    padding: 0;
+    margin: -1px;
+    overflow: hidden;
+    clip: rect(0, 0, 0, 0);
+    white-space: nowrap;
+    border-width: 0;
+  }
+
   .platform-modal {
     position: absolute;
-    top: 40px;
-    left: 60px;
+    top: calc(100% + 8px);
+    left: 0;
     min-width: 110px;
     background-color: #fff;
     border: 1px solid #ddd;
     border-radius: 4px;
     box-shadow: 0 2px 8px rgba(0, 0, 0, 0.15);
     z-index: 1001;
+    @include mobile {
+      min-width: 100px;
+    }
   }
 
   .platform-modal-content {
@@ -191,21 +217,16 @@
     align-items: center;
     width: 68px;
     padding: 0;
+    border: none;
+    background: none;
+    cursor: pointer;
   }
+
   .go-text {
     font-family: "Inter", sans-serif;
     font-weight: 800;
     font-size: 22px;
     letter-spacing: -0.5px;
-  }
-}
-
-@keyframes gradientMove {
-  0%,
-  100% {
-    background-position: 0% 50%;
-  }
-  50% {
-    background-position: 100% 50%;
+    color: #4270d3;
   }
 }
