+++ src/pages/Home.jsx
@@ -1,11 +1,10 @@
 import "./Home.scss";
 
-import Header from "../components/Header/Header";
-import HeroSection from "../components/HeroSection/HeroSection";
-import FeaturesSection from "../components/FeaturesSection/FeaturesSection";
-import Stats from "../components/Stats/Stats";
-import ToolSection from "../components/ToolSection/ToolSection";
-import Footer from "../components/Footer/Footer";
+import Header from "../components/common/Header/Header";
+import HeroSection from "../components/home/HeroSection/HeroSection";
+import FeaturesSection from "../components/home/FeaturesSection/FeaturesSection";
+import ToolSection from "../components/home/ToolSection/ToolSection";
+import Footer from "../components/common/Footer/Footer";
 
 const Home = () => {
   return (
@@ -18,7 +17,6 @@ const Home = () => {
       <div className="layout-container">
         <FeaturesSection />
       </div>
-      <Stats />
       <div className="footer-layout-container">
         <Footer />
       </div>
