+++ src/pages/Home.scss
@@ -1,3 +1,5 @@
+@use '../styles/breakpoints' as *;
+
 .home-container {
   display: flex;
   flex-direction: column;
@@ -5,19 +7,25 @@
   height: 100%;
   background-color: #1c1c1f;
 }
+
 .header-layout-container {
   display: flex;
   justify-content: center;
   width: 100%;
-  height: 40px;
+  height: 56px;
   font-weight: 700;
   background-color: #4270d3;
 }
+
 .layout-container {
   width: 1040px;
-  margin-bottom: 4px;
+  margin-bottom: 180px;
 
+  @include mobile {
+    width: 100%;
+  }
 }
+
 .footer-layout-container {
   display: flex;
   justify-content: center;
